refactor: Split steplist/utils into separate files.
Formerly, we were somewhere in between defining things in a monolithic index.ts file and having a separate file for each util. This goes fully towards separate files for each util, for consistency and build performance.
