+++ protocol-designer/src/steplist/formLevel/stepFormToArgs/test/mixFormToArgs.test.ts
@@ -8,15 +8,15 @@ import { fixture_96_plate } from '@opentrons/shared-data/labware/fixtures/2'
 import { AUTOMATIC } from '@opentrons/step-generation'
 
 import { DEFAULT_MM_BLOWOUT_OFFSET_FROM_TOP } from '/protocol-designer/constants'
+import { getOrderedWells } from '/protocol-designer/steplist/utils/getOrderedWells'
 
-import { getOrderedWells } from '../../../utils'
 import { mixFormToArgs } from '../mixFormToArgs'
 
 import type { LabwareDefinition2 } from '@opentrons/shared-data'
 import type { HydratedMixFormData } from '/protocol-designer/form-types'
 import type { GetCastFormData } from '/protocol-designer/steplist/fieldLevel'
 
-vi.mock('../../../utils')
+vi.mock('../../../utils/getOrderedWells')
 
 let castForm: GetCastFormData<HydratedMixFormData>
 const labwareDef = fixture_96_plate as LabwareDefinition2
