+++ protocol-designer/src/steplist/formLevel/stepFormToArgs/test/moveLiquidFormToArgs.test.ts
@@ -14,7 +14,7 @@ import {
   makeContext,
 } from '@opentrons/step-generation'
 
-import { getOrderedWells } from '../../../utils'
+import { getOrderedWells } from '../../../utils/getOrderedWells'
 import {
   getAirGapData,
   getMixData,
@@ -25,7 +25,7 @@ import type { LabwareDefinition2 } from '@opentrons/shared-data'
 import type { HydratedMoveLiquidFormData } from '/protocol-designer/form-types'
 import type { GetCastFormData } from '/protocol-designer/steplist/fieldLevel'
 
-vi.mock('../../../utils')
+vi.mock('../../../utils/getOrderedWells')
 vi.mock('assert')
 
 const ASPIRATE_WELL = 'A2' // default source is trough for these tests
